fix GetAvatar
tumblr's /avatar API:

a) tries to redirect to an image
b) provides the JSON in the initial response (before the redirect)
c) returns a 301 'Moved Permanently' in the initial response

so to get the JSON response, we check to see if the default client
has attempted a redirect. if so, we discard the redirect error and
attempt and use RoundTrip instead to fetch the JSON response.